[Build] Add abstract base plugin for module and docs plugins
